WIP Fix issue "updated_at" is not Ecto.DateTime
This was due to the fact that the incoming json data contains the field
"updated_at" and at runtime the `timestamps` macro gets expanded to
include the fields `updated_at` and `inserted_at` on the struct that is
used to parse the json data. This means the `updated_at` field gets
"hijacked" and instead of being a `Ecto.DateTime` type is instead the
string value that gets parsed from the data.